VP-1218: rename folder from converters to extensions
all concerned files have been moved
