+++ VirtoCommerce.JavaScriptShoppingCart.Data/Extensions/PaymentExtensions.cs
@@ -1,10 +1,11 @@
 using System.Linq;
-using VirtoCommerce.JavaScriptShoppingCart.Core.Model.Cart;
 using VirtoCommerce.JavaScriptShoppingCart.Core.Model.Common;
 using VirtoCommerce.Platform.Core.Common;
 using DomainCartModels = VirtoCommerce.Domain.Cart.Model;
+using Payment = VirtoCommerce.JavaScriptShoppingCart.Core.Model.Cart.Payment;
+using ShoppingCart = VirtoCommerce.JavaScriptShoppingCart.Core.Model.Cart.ShoppingCart;
 
-namespace VirtoCommerce.JavaScriptShoppingCart.Data.Converters
+namespace VirtoCommerce.JavaScriptShoppingCart.Data.Extensions
 {
     public static class PaymentExtensions
     {
