+++ VirtoCommerce.JavaScriptShoppingCart.Data/VirtoCommerce.JavaScriptShoppingCart.Data.csproj
@@ -52,13 +52,13 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Converters\AddressExtensions.cs" />
-    <Compile Include="Converters\CartConverter.cs" />
-    <Compile Include="Converters\DiscountExtensions.cs" />
-    <Compile Include="Converters\PaymentExtensions.cs" />
-    <Compile Include="Converters\ShipmentExtensions.cs" />
-    <Compile Include="Converters\ShipmentItemExtensions.cs" />
-    <Compile Include="Converters\TaxDetailExtensions.cs" />
+    <Compile Include="Extensions\AddressExtensions.cs" />
+    <Compile Include="Extensions\ShoppingCartExtensions.cs" />
+    <Compile Include="Extensions\DiscountExtensions.cs" />
+    <Compile Include="Extensions\PaymentExtensions.cs" />
+    <Compile Include="Extensions\ShipmentExtensions.cs" />
+    <Compile Include="Extensions\ShipmentItemExtensions.cs" />
+    <Compile Include="Extensions\TaxDetailExtensions.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Services\CartBuilder.cs" />
   </ItemGroup>
